backend/requirements.txt
fastapi
sqlalchemy
passlib[bcrypt]
pytest
uvicorn
alembic
PyJWT
